A test showing that the distribution isn't very good yet
This module isn't useful unless it'll typically produce a reasonably good
spread of guests over hosts. This test shows we haven't achieved that:
two of the hosts here are very close on the circle and so a1 ends up
having considerably less chance of having any particular guest assigned
to it.